7-Regular Expressions/numb3rs/test_numb3rs.py
# Import the 'validate' function from the 'numb3rs' module.
from numb3rs import validate


def main():
    test_validate()  # Run the test function.


def test_validate():
    # Test cases for the 'validate' function.
    assert validate("127.0.0.1") == True  # Valid IP address
    assert validate("255.255.255.255") == True  # Valid IP address
    # Invalid IP address (out of range)
    assert validate("512.512.512.512") == False
    assert validate("1.2.3.1000") == False  # Invalid IP address (out of range)
    assert validate("cat") == False  # Invalid input (not an IP address)
    # Invalid IP address (out of range)
    assert validate("75.456.76.65") == False
